Remove get_field() function
The $ or [[ subsetting operators should be used instead of get_field(). get_field() was originally intended to output NA if a field didn't exist. However given all feeds should follow the GTFS-R specification and all fields are present (even if a NULL value is given), this was not an issue. Unfortunately the get_field() function persisted, even when it was modified to not return an NA value.

realtime_X() functions have been rewritten to use the '$' operator.

Documentation has been updated accordingly.

TripDescriptor fields in Alert feeds are now included (the documentation was misinterpreted).

Increment version to 0.2.0